requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-db-logger==0.1.12
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-api-tracking==1.8.0
drf-tracking==1.5.0
environ==1.0
idna==3.4
importlib-metadata==5.0.0
JSON-log-formatter==0.5.1
Markdown==3.4.1
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
urllib3==1.26.13
wincertstore==0.2
zipp==3.10.0
